requirements.txt
setuptools
annotated-types==0.7.0
asttokens==3.0.0
backcall==0.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
debugpy==1.8.10
decorator==5.1.1
docker-pycreds==0.4.0
eval-type-backport==0.2.0
executing==2.1.0
filelock==3.16.1
fsspec==2024.10.0
gitdb==4.0.11
gitpython==3.1.43
idna==3.10
importlib-metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
jinja2==3.1.4
jupyter-client==8.6.3
jupyter-core==5.7.2
markupsafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
prompt-toolkit==3.0.48
protobuf==5.29.1
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.10.3
pydantic-core==2.27.1
pygments==2.18.0
python-dateutil==2.9.0.post0
pyyaml==6.0.2
pyzmq==26.2.0
requests==2.32.3
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.17.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.13.3
torch==2.4.1
torchaudio==2.4.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.0.0
typing-extensions==4.12.2
urllib3==2.2.3
wandb==0.19.0
wcwidth==0.2.13
zipp==3.20.2

